chore(snapengage): move sdk loader to a separate file
