+++ src/integrations/SnapEngage/browser.js
@@ -8,6 +8,7 @@ import { recordingLiveChatEvents } from './util';
 import { getHashFromArray } from '../../utils/commonUtils';
 import { isObject } from '../../utils/utils';
 import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
+import { loadNativeSdk } from './nativeSdkLoader';
 
 class SnapEngage {
   constructor(config, analytics, destinationInfo) {
@@ -26,24 +27,7 @@ class SnapEngage {
   }
 
   loadScript() {
-    (function (widgetId) {
-      const se = document.createElement('script');
-      se.type = 'text/javascript';
-      se.async = true;
-      se.src = `https://storage.googleapis.com/code.snapengage.com/js/${widgetId}.js`;
-      se.setAttribute('data-loader', LOAD_ORIGIN);
-      let done = false;
-      se.onload = se.onreadystatechange = function () {
-        if (
-          !done &&
-          (!this.readyState || this.readyState === 'loaded' || this.readyState === 'complete')
-        ) {
-          done = true;
-        }
-      };
-      const s = document.getElementsByTagName('script')[0];
-      s.parentNode.insertBefore(se, s);
-    })(this.widgetId);
+    loadNativeSdk(this.widgetId);
   }
 
   init() {
